añadida opciones de cuenta al admin, añadida la opcion de gestionar reservas al hotel, cambiado el modo del que se accede a la direccion del cliente, ahora lo hago por la relacion la cual tambien he tenido que cambiar, añadida validacion para valores decimales al precio de habitacion, añadido el tiempo en el que se publicó la habitacion, cambiado el archivo JS que calcula el precio de la pension, ahora lo calcula con decimales

+++ public/js/pago/sumprice.js
@@ -13,8 +13,8 @@ pensiones.onchange = changePrice;
 
 async function changePrice(event){
     let precioExtra = await requestPrice(event.target.value);
-    precio.textContent = parseInt(precioOriginal) + parseInt(precioExtra);
-    precio_form.value = parseInt(precioOriginal) + parseInt(precioExtra);
+    precio.textContent = parseFloat(precioOriginal) + parseFloat(precioExtra);
+    precio_form.value = parseFloat(precioOriginal) + parseFloat(precioExtra);
     precioExtras.textContent = precioExtra;
 
     pension_form.value = event.target.value;
